feat: Alpha 1 + ElectronPlayer example app alpha
This is the first working version of NativeKit. It uses JavaScriptCore to run a js file embedded inside a template app. This JS then uses an API exposed from swift to create an app. The app's main window is just a WKWebView browser which has a special event emitter system which allows it to communicate with the JavaScriptCore backend. The ElectronPlayer app is a proof of concept based on the old Electron based app.