feat: Typescript support + default menu in template app
Added support for Typescript with the 'lib/' package. Also added 'tsconfig.json' to project to enforce strict typechecking. Added "Always On Top" option to BrowserWindow and ElectronPlayer example app. Restructured packages so that the cli and lib are together to make publishing to npm in the future easier. Added better error handling and better fullscreen support to the WKWebView (more work still needs to be done). Moved the Window controller to a new file for the planned multiwindow support in the future. Added Apple App Sandbox to the Info.plist file. Support for checkboxes in the Menu template.